requirements.txt
aiogram>=3.14,<4
python-dotenv>=1.0.1,<2
PyYAML>=6.0.2,<7
